requirements.txt
flask
pygame
pillow
Flask-SQLAlchemy
psycopg2